refactor: extract out the intialisation code
