+++ examples/expo-example/components/ControlExamples/Boolean/Boolean.stories.tsx
@@ -1,4 +1,4 @@
-import type { ComponentMeta, ComponentStoryObj } from '@storybook/react-native';
+import type { ComponentMeta, ComponentStoryObj } from '@storybook/react-native/types';
 import { Switch } from './Boolean';
 
 const BooleanExample: ComponentMeta<typeof Switch> = {
@@ -11,3 +11,5 @@ export default BooleanExample;
 type BooleanStory = ComponentStoryObj<typeof Switch>;
 
 export const Basic: BooleanStory = {};
+
+export const On: BooleanStory = { args: { on: true } };
