+++ packages/react-native/theme.ts
@@ -1,0 +1,1 @@
+export { darkTheme, theme, type Theme } from '@storybook/react-native-theming';
