+++ packages/react-native/tsup.config.ts
@@ -2,12 +2,12 @@ import { defineConfig } from 'tsup';
 
 export default defineConfig((options) => {
   return {
-    entry: ['src/index.ts'],
+    entry: ['src/index.ts', 'src/preview/StartV7.tsx'],
     // minify: !options.watch,
     clean: !options.watch,
     dts: !options.watch
       ? {
-          entry: ['src/index.ts'],
+          entry: ['src/index.ts', 'src/preview/StartV7.tsx'],
           resolve: true,
         }
       : false,
