+++ packages/react-native/types.ts
@@ -1,0 +1,1 @@
+export * from './src/types/types-6.0';
